api_requirements.txt
# API Requirements
# Install with: pip install -r api_requirements.txt

# Core FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
aiofiles==24.1.0

# Pydantic
pydantic==2.9.2
pydantic-settings==2.6.0

# Document processing
python-docx==1.1.2
pdf2image==1.17.0
reportlab==4.2.5
Pillow==10.4.0

# PyTorch (choose your version based on CPU/GPU)
# For CPU only:
# torch==2.4.0 torchvision==0.19.0 --index-url https://download.pytorch.org/whl/cpu

# For CUDA 11.8:
# torch==2.4.0 torchvision==0.19.0 --index-url https://download.pytorch.org/whl/cu118

# For CUDA 12.1:
# torch==2.4.0 torchvision==0.19.0 --index-url https://download.pytorch.org/whl/cu121

# Optional: Redis for task queue (if needed in future)
# redis==5.0.1
